Added in some more instructions/commands that are useful to remember into README.md.
Added in a UsersController that will handle all user operations.

Added in basic preliminary routes to handle how accessing the website
will occur.

Added preliminary views. TODO: Controllers need to be written to signin,
authenticate, signout, and show the user's profile.

Also TODO: Create a profile table to house some more information about the
user. Create a new migration, add a profile table with a foreign key
back to the users table. Also add a foreign key to the users table to
point to the profile, on update cascade and on delete cascade.