Finally have a proper and basic authentication system in place. The system properly registers a new user. User creation is now within a Eloquent transaction, so the action is ATOMIC.
Changed the filters to ensure that a logged in user cannot
relogin and reregister until they are logged out. The routes
make use of this new filter.

Also, I have a profile referencing (with a FK) a user. The DB
migration simply states that a profile points to a user.

Then the relationship is specified within the User and Profile
objects. For instance, even though we did not specify a FK to
a profile in the users table, the fact that the profile has a FK
to their user, coupled with the relationship stated in Eloquent,
operations understand they are linked together.